Fix smart contract address negotiation and reorg tests #1499 (#1516)
* WIP. Improved smart contract negotiation. Reorganizing tests.

* Docker tests reorg WIP.

* Reorg finished. Some tests fail.

* Fix module name and clippy.

* WIP. Updated toolchain. Clippy fails.

* Fix clippy.

* Fix docker tests compilation.

* Use cargo rustc on building mobile libs.

* Fix tests.

* Fix now_ms import.

* WIP.

* WIP.

* WIP.

* WIP. Porting WASM integration tests from mm2_tests to a separate module.

* WIP. Ported test_qrc20_tx_history to wasm_tests.

* Turn off doctest for all workspace crates.

* Move structs to mm2_test_helpers.

* Add tests.

* Fix clippy.

* Fix review notes.

* Fix one more review note.

+++ mm2src/coins/lightning/ln_sql.rs
@@ -563,8 +563,8 @@ impl LightningDB for SqliteLightningDB {
 
     async fn add_channel_to_db(&self, details: DBChannelDetails) -> Result<(), Self::Error> {
         let for_coin = self.db_ticker.clone();
-        let rpc_id = details.rpc_id as i64;
-        let created_at = details.created_at as i64;
+        let rpc_id = details.rpc_id;
+        let created_at = details.created_at;
 
         let sqlite_connection = self.sqlite_connection.clone();
         async_blocking(move || {
